// shopify/utils/storefront.graphql.gen.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K];
};
// ...
export type Scalars = {
  ID: { input: string; output: string };
  String: { input: string; output: string };
  Boolean: { input: boolean; output: boolean };
  Int: { input: number; output: number };
  Float: { input: number; output: number };
  Color: { input: any; output: any };
  DateTime: { input: any; output: any };
  Decimal: { input: any; output: any };
  HTML: { input: any; output: any };
  JSON: { input: any; output: any };
  URL: { input: any; output: any };
  UnsignedInt64: { input: any; output: any };
};
// ...
export type MediaContentType =
  /** An externally hosted video. */
  | "EXTERNAL_VIDEO"
  /** A Shopify hosted image. */
  | "IMAGE"
  /** A 3d model. */
  | "MODEL_3D"
  /** A Shopify hosted video. */
  | "VIDEO";
// ...
export type UnitPriceMeasurementMeasuredType =
  /** Unit of measurements representing areas. */
  | "AREA"
  /** Unit of measurements representing lengths. */
  | "LENGTH"
  /** Unit of measurements representing volumes. */
  | "VOLUME"
  /** Unit of measurements representing weights. */
  | "WEIGHT";

/** The valid units of measurement for a unit price measurement. */
export type UnitPriceMeasurementMeasuredUnit =
  /** 100 centiliters equals 1 liter. */
  | "CL"
  /** 100 centimeters equals 1 meter. */
  | "CM"
  /** Metric system unit of weight. */
  | "G"
  /** 1 kilogram equals 1000 grams. */
  | "KG"
  /** Metric system unit of volume. */
  | "L"
  /** Metric system unit of length. */
  | "M"
  /** Metric system unit of area. */
  | "M2"
  /** 1 cubic meter equals 1000 liters. */
  | "M3"
  /** 1000 milligrams equals 1 gram. */
  | "MG"
  /** 1000 milliliters equals 1 liter. */
  | "ML"
  /** 1000 millimeters equals 1 meter. */
  | "MM";
// ...
export type WeightUnit =
  /** Metric system unit of mass. */
  | "GRAMS"
  /** 1 kilogram equals 1000 grams. */
  | "KILOGRAMS"
  /** Imperial system unit of mass. */
  | "OUNCES"
  /** 1 pound equals 16 ounces. */
  | "POUNDS";
// ...
export type SearchProductsQueryVariables = Exact<{
  first?: InputMaybe<Scalars["Int"]["input"]>;
  after?: InputMaybe<Scalars["String"]["input"]>;
  query?: InputMaybe<Scalars["String"]["input"]>;
}>;

export type SearchProductsQuery = {
  products: {
    pageInfo: { hasNextPage: boolean };
    nodes: Array<
      {
        availableForSale: boolean;
        createdAt: any;
        description: string;
        descriptionHtml: any;
        handle: string;
        id: string;
        isGiftCard: boolean;
        onlineStoreUrl?: any | null;
        productType: string;
        publishedAt: any;
        requiresSellingPlan: boolean;
        tags: Array<string>;
        title: string;
        totalInventory?: number | null;
        updatedAt: any;
        vendor: string;
        featuredImage?: { altText?: string | null; url: any } | null;
        images: { nodes: Array<{ altText?: string | null; url: any }> };
        media: {
          nodes: Array<
            {
              alt?: string | null;
              mediaContentType: MediaContentType;
              previewImage?: { altText?: string | null; url: any } | null;
            } | {
              alt?: string | null;
              mediaContentType: MediaContentType;
              previewImage?: { altText?: string | null; url: any } | null;
            } | {
              alt?: string | null;
              mediaContentType: MediaContentType;
              previewImage?: { altText?: string | null; url: any } | null;
            } | {
              alt?: string | null;
              mediaContentType: MediaContentType;
              previewImage?: { altText?: string | null; url: any } | null;
            }
          >;
        };
        options: Array<{ name: string; values: Array<string> }>;
        priceRange: {
          minVariantPrice: { amount: any; currencyCode: CurrencyCode };
          maxVariantPrice: { amount: any; currencyCode: CurrencyCode };
        };
        seo: { title?: string | null; description?: string | null };
        variants: {
          nodes: Array<
            {
              availableForSale: boolean;
              barcode?: string | null;
              currentlyNotInStock: boolean;
              id: string;
              quantityAvailable?: number | null;
              requiresShipping: boolean;
              sku?: string | null;
              title: string;
              weight?: number | null;
              weightUnit: WeightUnit;
              compareAtPrice?:
                | { amount: any; currencyCode: CurrencyCode }
                | null;
              image?: { altText?: string | null; url: any } | null;
              price: { amount: any; currencyCode: CurrencyCode };
              selectedOptions: Array<{ name: string; value: string }>;
              unitPrice?: { amount: any; currencyCode: CurrencyCode } | null;
              unitPriceMeasurement?: {
                measuredType?: UnitPriceMeasurementMeasuredType | null;
                quantityValue: number;
                referenceUnit?: UnitPriceMeasurementMeasuredUnit | null;
                quantityUnit?: UnitPriceMeasurementMeasuredUnit | null;
              } | null;
            }
          >;
        };
      }
    >;
  };
};